feature(batch action) add batch acion start domain command and handler
