+++ module/batch-action/src/Infrastructure/Persistence/Repository/Mapper/DbalBatchActionMapper.php
@@ -10,6 +10,9 @@ declare(strict_types=1);
 namespace Ergonode\BatchAction\Infrastructure\Persistence\Repository\Mapper;
 
 use Ergonode\BatchAction\Domain\Entity\BatchAction;
+use Ergonode\BatchAction\Domain\Entity\BatchActionId;
+use Ergonode\BatchAction\Domain\ValueObject\BatchActionType;
+use Ergonode\BatchAction\Domain\ValueObject\BatchActionAction;
 
 class DbalBatchActionMapper
 {
@@ -23,5 +26,29 @@ class DbalBatchActionMapper
             'resource_type' => $batchAction->getType()->getValue(),
             'action' => $batchAction->getAction()->getValue(),
         ];
+    }
+
+    /**
+     * @throws \ReflectionException
+     */
+    public function create(array $record): BatchAction
+    {
+        $reflector = new \ReflectionClass(BatchAction::class);
+        /** @var BatchAction $object */
+        $object = $reflector->newInstanceWithoutConstructor();
+
+        $map = [
+            'id' => new BatchActionId($record['id']),
+            'type' => new BatchActionType($record['resource_type']),
+            'action' => new BatchActionAction($record['action']),
+        ];
+
+        foreach ($map as $key => $value) {
+            $property = $reflector->getProperty($key);
+            $property->setAccessible(true);
+            $property->setValue($object, $value);
+        }
+
+        return $object;
     }
 }
